win32 platform support checks

+++ glue/glue.py
@@ -1,3 +1,4 @@
+import subprocess
 import sys
 import os
 import signal
@@ -12,6 +13,11 @@ LOG_COLOR = "\033[38;5;69m"
 MSG_COLOR = "\033[38;5;195m"
 ERROR_COLOR = "\033[38;5;1m"
 TITLE = "\033[1m"
+
+def is_win32():
+    if os.name.lower() == "nt":
+        return True
+    return False
 
 async def process_stdout(proc, name):
     try:
@@ -51,19 +57,34 @@ async def pretty_stderr(proc, name):
 
 
 async def start_component(cmd, name):
+    is_win = is_win32()
+
     print(f"{TITLE}Starting {name}{RESET}")
     name_b = name.encode("utf-8")
 
     msg_channel = ch.create_channel(100)
     component_channels[name_b] = msg_channel
 
-    proc = await asyncio.create_subprocess_shell(
-        cmd,
-        cwd=os.path.join("components", name), # Undocumented!
-        stdin=asyncio.subprocess.PIPE,
-        stdout=asyncio.subprocess.PIPE,
-        stderr=asyncio.subprocess.PIPE,
-    )
+    proc = None
+
+    if is_win:
+        print(f"{LOG_COLOR}LOG [{name}]{RESET} => Win32 platform detected, using: CREATE_NEW_PROCESS_GROUP flag...={name}")
+        proc = await asyncio.create_subprocess_shell(
+            cmd,
+            cwd=os.path.join("components", name), # Undocumented!
+            stdin=asyncio.subprocess.PIPE,
+            stdout=asyncio.subprocess.PIPE,
+            stderr=asyncio.subprocess.PIPE,
+            creationflags=subprocess.CREATE_NEW_PROCESS_GROUP
+        )
+    else:
+        proc = await asyncio.create_subprocess_shell(
+            cmd,
+            cwd=os.path.join("components", name), # Undocumented!
+            stdin=asyncio.subprocess.PIPE,
+            stdout=asyncio.subprocess.PIPE,
+            stderr=asyncio.subprocess.PIPE
+        )
 
     stdout_processor = asyncio.create_task(process_stdout(proc, name_b))
     stdin_processor = asyncio.create_task(send_stdin(proc, msg_channel, name))
@@ -80,19 +101,33 @@ async def start_component(cmd, name):
     stderr_logger.cancel()
 
 async def main():
-    os.setpgrp() # create new process group, become its leader
+    is_win = is_win32()
+    tty = "dev/ttys005"
+    py_cmd = "python3"
 
+    if not is_win:
+        os.setpgrp() # create new process group, become its leader
+    
     try:
+        if is_win:
+            print("Win32 platform detected!!!")
+            tty = "/dev/cons1"
+            py_cmd = "python"
+
         await asyncio.gather(
-            start_component("python3 run.py /dev/ttys005", "INTR"),
-            start_component("python3 run.py", "RLAY"),
-            start_component("python3 run.py", "CONS"),
+            start_component(f"{py_cmd} run.py {tty}", "INTR"),
+            start_component(f"{py_cmd} run.py", "RLAY"),
+            start_component(f"{py_cmd} run.py", "CONS"),
         )
     except:
         traceback.print_exc()
     finally:
         print("bye")
-        os.killpg(0, signal.SIGKILL) # kill all processes in my group
+        
+        if not is_win:
+            os.killpg(0, signal.SIGKILL) # kill all processes in my group
+
+        os.kill(0, signal.SIGTERM)
 
 if __name__ == "__main__":
     asyncio.run(main())
